Rewrite ci logic
