+++ seatunnel-e2e/seatunnel-connector-v2-e2e/connector-mongodb-e2e/src/test/resources/cdcIT/fake_cdc_sink_mongodb.conf
@@ -19,20 +19,57 @@
 ######
 
 env {
-  # You can set engine configuration here
   execution.parallelism = 1
-  job.mode = "STREAMING"
-  execution.checkpoint.interval = 5000
+  job.mode = "BATCH"
 }
 
 source {
-  MySQL-CDC {
-    result_table_name = "customers_mysql_cdc"
-    server-id = 5652
-    username = "st_user"
-    password = "seatunnel"
-    table-names = ["mysql_cdc.products"]
-    base-url = "jdbc:mysql://mysql_cdc_e2e:3306/mysql_cdc"
+  FakeSource {
+    schema = {
+      fields {
+        c_int = bigint
+        name = string
+        score = int
+      }
+    }
+    rows = [
+      {
+        kind = INSERT
+        fields = [1, "A", 100]
+      },
+      {
+        kind = INSERT
+        fields = [2, "B", 100]
+      },
+      {
+        kind = INSERT
+        fields = [3, "C", 100]
+      },
+      {
+        kind = INSERT
+        fields = [3, "C", 100]
+      },
+      {
+        kind = INSERT
+        fields = [3, "C", 100]
+      },
+      {
+        kind = INSERT
+        fields = [3, "C", 100]
+      }
+      {
+        kind = UPDATE_BEFORE
+        fields = [1, "A", 100]
+      },
+      {
+        kind = UPDATE_AFTER
+        fields = [1, "A_1", 100]
+      },
+      {
+        kind = DELETE
+        fields = [2, "B", 100]
+      }
+    ]
   }
 }
 
@@ -47,8 +84,7 @@ sink {
       fields {
         c_int = int
         name = string
-        description = string
-        weight = float
+        score = int
       }
     }
   }
