--- seatunnel-e2e/seatunnel-connector-v2-e2e/connector-mongodb-e2e/src/test/resources/ddl/mysql_cdc.sql
@@ -1,42 +1,0 @@
---
--- Licensed to the Apache Software Foundation (ASF) under one or more
--- contributor license agreements.  See the NOTICE file distributed with
--- this work for additional information regarding copyright ownership.
--- The ASF licenses this file to You under the Apache License, Version 2.0
--- (the "License"); you may not use this file except in compliance with
--- the License.  You may obtain a copy of the License at
---
---     http://www.apache.org/licenses/LICENSE-2.0
---
--- Unless required by applicable law or agreed to in writing, software
--- distributed under the License is distributed on an "AS IS" BASIS,
--- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
--- See the License for the specific language governing permissions and
--- limitations under the License.
---
-
--- ----------------------------------------------------------------------------------------------------------------
--- DATABASE:  inventory
--- ----------------------------------------------------------------------------------------------------------------
-CREATE DATABASE IF NOT EXISTS `mysql_cdc`;
-
-use mysql_cdc;
--- Create a mysql data source table
-CREATE TABLE products (
-                          c_int INTEGER NOT NULL AUTO_INCREMENT PRIMARY KEY,
-                          name VARCHAR(255) NOT NULL DEFAULT 'SeaTunnel',
-                          description VARCHAR(512),
-                          weight INTEGER
-);
-ALTER TABLE products AUTO_INCREMENT = 101;
-
-INSERT INTO products
-VALUES (default,"scooter","Small 2-wheel scooter",3),
-       (default,"car battery","12V car battery",8),
-       (default,"12-pack drill bits","12-pack of drill bits with sizes ranging from #40 to #3",8),
-       (default,"hammer","12oz carpenter's hammer",2),
-       (default,"hammer","14oz carpenter's hammer",5),
-       (default,"hammer","16oz carpenter's hammer",6),
-       (default,"rocks","box of assorted rocks",7),
-       (default,"jacket","water resistent black wind breaker",1),
-       (default,"spare tire","24 inch spare tire",22);
